requirements.txt
# requirements.txt
pandas
python-chess
apache-beam
pytest
pytest-cov